Peptide-Properties-Project/requirements.txt
torch
biopython
numpy
scikit-learn
tqdm
